Chart name must be unique per version
This changes the `metadata.name` attribute on the Chart.yaml file to be
unique per chart version. This is an unfortunate burden of making it
possible to install the Chart.yaml into a cluster, and have the name be
unique per thing we care about.

h/t to @kfox1111 for noticing this.