Change the Lua sandbox model for charts.
This is a result of some work on @prydonius's question about library
charts and Lua. The change here is that there is only one Lua
interpreter for a chart operation. So if a parent chart has three
subcharts, there is only one Lua interpreter and it has access to all
four charts.

This change allows two things:

- library charts with Lua
- explicit override control for a top-level chart