Broaden the scope of a helm subproject to include non-tech projects
The existing wording focused on technical codebases and did not
appear to provide room for a project dealing with community
management or any other non-technical project.

This change opens the door for projects that are not technical in
nature. It is added at this point to support community management
which has happened for a long time but has not been recognized or
empowered the way other projects have been. I hope that recognizing
the work and officially empowering it helps to expand the effort
and grow help usage and contributors.

Note, one of the lines in the maintainers structure section just
reiterated what was said in the paragraph it was apart of. That
line was removed due to redundancy.

Signed-off-by: Matt Farina <matt@mattfarina.com>