Removing initial election details from governance
When the first Helm Org Maintainers election occured there were
criteria laid out to ensure sub-project diversity at that time.
This section of the governance was only for that initial election
and did not apply after the initial election.

As that election was years ago and the governance is stored in
git, this PR removes this section as it is no longer relevant.

Signed-off-by: Matt Farina <matt@mattfarina.com>