HIP Support Helm release data stored in multiple K8s Secrets
Helm release data encoded into the release field in a Kubernetes Secret can cause the Secret as a whole to exceed 1 MiB (as defined by MaxSecretSize in kubernetes/pkg/apis/core/types.go). When this happens, the Kubernetes Secret cannot be stored in Etcd and the Helm install or upgrade fails. Splitting the Helm release data across multiple Kubernetes Secrets resolves this problem.

Signed-off-by: Ralf Van Dorsselaer <ralfvandorsselaer@gmail.com>